+++ packages/components/src/popover/Placement.js
@@ -2,6 +2,8 @@ const Placement = {
   TOP: 'top',
   RIGHT: 'right',
   BOTTOM: 'bottom',
+  BOTTOM_RIGHT: 'bottom-right',
+  BOTTOM_LEFT: 'bottom-left',
   LEFT: 'left',
   LEFT_TOP: 'left-top',
   RIGHT_TOP: 'right-top',