+++ packages/components/src/popover/positioning.spec.js
@@ -7,14 +7,27 @@ describe('Popover positioning', () => {
       expect(getPlacement(undefined, Placement.LEFT)).toBe('');
     });
 
-    it('is corrected to bottom if overflows left and right', () => {
-      documentWidth(320);
-      const popover = new Popover({ width: 150 }).trigger({
-        left: 20,
-        width: 200,
+    describe('is corrected to bottom if overflows left and right', () => {
+      it('corrects to bottom center if it doesnt overflow in this position', () => {
+        documentWidth(360);
+        const popover = new Popover({ width: 360 }).trigger({ left: 130, width: 100 });
+
+        expect(getPlacement(popover, Placement.LEFT)).toBe(Placement.BOTTOM);
       });
 
-      expect(getPlacement(popover, Placement.LEFT)).toBe(Placement.BOTTOM);
+      it('corrects to bottom right if it overflows left when positioned bottom', () => {
+        documentWidth(360);
+        const popover = new Popover({ width: 360 }).trigger({ left: 129, width: 100 });
+
+        expect(getPlacement(popover, Placement.LEFT)).toBe(Placement.BOTTOM_RIGHT);
+      });
+
+      it('corrects to bottom left if it overflows right when positioned bottom', () => {
+        documentWidth(360);
+        const popover = new Popover({ width: 360 }).trigger({ left: 131, width: 100 });
+
+        expect(getPlacement(popover, Placement.LEFT)).toBe(Placement.BOTTOM_LEFT);
+      });
     });
 
     it('is corrected to left if right and overflows right', () => {
@@ -82,12 +95,13 @@ describe('Popover positioning', () => {
 
       expect(getPopoverPosition(popover, Placement.TOP)).toEqual({
         top: 250,
+        right: 450,
         left: 350,
       });
     });
 
     it('is correct for right placement', () => {
-      const popover = new Popover({ height: 50 }).trigger({
+      const popover = new Popover({ width: 100, height: 50 }).trigger({
         top: 300,
         left: 300,
         width: 200,
@@ -96,6 +110,7 @@ describe('Popover positioning', () => {
 
       expect(getPopoverPosition(popover, Placement.RIGHT)).toEqual({
         top: 375,
+        right: 600,
         left: 500,
       });
     });
@@ -110,7 +125,42 @@ describe('Popover positioning', () => {
 
       expect(getPopoverPosition(popover, Placement.BOTTOM)).toEqual({
         top: 500,
+        right: 450,
         left: 350,
+      });
+    });
+
+    it('is correct for bottom left placement', () => {
+      const popover = new Popover({ width: 100 })
+        .trigger({
+          top: 300,
+          left: 300,
+          width: 200,
+          height: 200,
+        })
+        .arrow({ width: 14, right: 29, marginLeft: -7 });
+
+      expect(getPopoverPosition(popover, Placement.BOTTOM_LEFT)).toEqual({
+        top: 500,
+        right: 429,
+        left: 329,
+      });
+    });
+
+    it('is correct for bottom right placement', () => {
+      const popover = new Popover({ width: 100 })
+        .trigger({
+          top: 300,
+          left: 300,
+          width: 200,
+          height: 200,
+        })
+        .arrow({ width: 14, left: 29, marginLeft: -7 });
+
+      expect(getPopoverPosition(popover, Placement.BOTTOM_RIGHT)).toEqual({
+        top: 500,
+        right: 471,
+        left: 371,
       });
     });
 
@@ -123,17 +173,19 @@ describe('Popover positioning', () => {
 
       expect(getPopoverPosition(popover, Placement.LEFT)).toEqual({
         top: 375,
+        right: 300,
         left: 200,
       });
     });
 
     it('is correct for top right placement', () => {
-      const popover = new Popover()
+      const popover = new Popover({ width: 100 })
         .trigger({ top: 300, left: 300, width: 200, height: 200 })
         .arrow({ top: 20, height: 10, marginTop: 8 });
 
       expect(getPopoverPosition(popover, Placement.RIGHT_TOP)).toEqual({
         top: 367,
+        right: 600,
         left: 500,
       });
     });
@@ -145,6 +197,7 @@ describe('Popover positioning', () => {
 
       expect(getPopoverPosition(popover, Placement.LEFT_TOP)).toEqual({
         top: 367,
+        right: 300,
         left: 200,
       });
     });
@@ -189,9 +242,13 @@ class Popover {
 
   arrow = styles => {
     window.getComputedStyle = jest.fn().mockReturnValue({
-      top: styles.top.toString(),
-      height: styles.height.toString(),
-      'margin-top': styles.marginTop.toString(),
+      top: styles.top ? styles.top.toString() : '',
+      right: styles.right ? styles.right.toString() : '',
+      left: styles.left ? styles.left.toString() : '',
+      height: styles.height ? styles.height.toString() : '',
+      width: styles.width ? styles.width.toString() : '',
+      'margin-top': styles.marginTop ? styles.marginTop.toString() : '',
+      'margin-left': styles.marginLeft ? styles.marginLeft.toString() : '',
     });
 
     return this;